[NEW] Changed format of exported experiment file.
Experiment files generated by the plugin now list each measure in a
separate line, along with all its parameters (and no blank lines in
between). This is the same format that is recognised by the CL version.
Also fixed a bug with parsing lists of doubles in the new measurement
dialog.