[NEW] Various enhancements to command line version
Several small additions to the command line interface:

- Progress monitoring: a message will be printed every time 10%, 20%,
... of the total simulation time (counting all replications) has been
reached. If running multithreaded, this applies to each subtask.
- Parameter overriding: aspects of the simulation can be configured
through the -deadline and -replications option, effectively overriding
the final time and number of replications specified in the experiment
file. The -seed option lets the user set the initial seed of the random
number generator for the simulations.
- Richer output: in addition to the measure files, the output directory
now contains a copy of the original model, a new experiment file, a file
with the execution time of the simulations ("timeInSeconds") and an
"info" file with a summary of the conditions of the simulation.
Execution time is also reported in standard output.

A class for MultiVeStA experiments has also been added, although the
full functionality is commented out.