[FIX] Complete CarmaModelToCode
CarmaModelToCode was missing some functionality, including the
representation of the space model. This has now been added and should
cover all cases. Other changes include fixing a missing parenthesis in
the single-argument uniform (U) function, and the representation of
atomic real constants. The latter was changed to avoid scientific
notation, which is not recognised by the tool. The current state could
still be improved, as it uses BigDecimal when necessary, which may print
unexpectedly many digits.