[NEW] Allow scientific notation for real numbers
Real numbers were previously only recognised in decimal notation in CaSL
files. This is now extended to allow for scientific notation of the form
1.0e-5. Note that the decimal separator is required, and that the
exponent is indicated only by lower-case e, to avoid confusion with
(Euler's constant) E. This allows us to simplify the printing of atomic
real values when generating CaSL code, removing the need for BigDecimal.