[NEW] Add validators for updates and assignments to collections
The new syntax for assigning to and updating a list element required new
checks, which have now been added as validators (eg that indices are
integers or that the right type of element is assigned). This change
also includes checks for the update of adding or removing a collection
element, and minor fixes in CarmaModelToCode.