[FIX] Ensure sequential order of updates in actions
Previously, to update a component's store after an action occurs, all
attribute values were retrieved right before the update occurred. This
could lead to unintuitive behaviour if an attribute was updated based on
the value of another previously-updated attribute. This change ensures
that the most recent value of each attribute is used. It also fixes some
small bugs with validation checks and Java code generation.