[NEW] Add the Step (Interactive Execution) View
A new view is added to the plugin, allowing the user to explore the
model's behaviour step-by-step, by choosing one action at a time and
observing its effects. At the moment the interface is fairly basic, and
there is scope for adding more selection and filtering options for the
user, as well as improving the display of the model state and history.