Add weighted random choice
This update introduces a new syntactic construct for randomly selecting
from a list of elements based on user-specified weights. This could
previously be accomplished with the select function, but the new syntax
lets the user explicitly specify the weights if they are known
beforehand, which should save some computations. The new syntax is:
selectFrom(e1 : w1, e2: w2, ... , en : wn)
where e1, ..., en are expressions of the same type, and w1, ..., wn are
real-valued expressions. The whole expression returns one of the ei's,
with the i-th element being chosen with probability wi / (w1 + ... wn).